// generated/src/app/api-models/cap-cptdt-rpt-wrk.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapCptdtRptWrk {
  seqCapCptdtRptWrk: number;
  seqJobId: number;
  capModelId: string;
  capEntityCode: string;
  capEntityId: string;
  capPoolId: string;
  seqCapPoolId: number;
  provLastName: string;
  provFirstName: string;
  vendFullName: string;
  applyTo: string;
  capStoplossId: string;
  capMonth: Date;
  capMonthRun: Date;
  capTransAmt: number;
  pmpmWitholdAmt: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  adjustmentFlag: string;
  seqIncentiveRule: number;
  seqProvContract: number;
  accessIncentiveAmount: number;
  accessIncentivePercent: number;
  accessCapApply: string;
  incentiveRetro: string;
}